refactor(bienbot-ui): extract NotificationItem and ImageAttachment
Move self-contained NotificationItem and ImageAttachment sub-components
out of BienBotPanel.jsx into their own files to reduce module size.
